requirements.txt
pydantic-ai==0.1.3
pydantic-ai-slim[tavily]==0.1.3
streamlit
python-dotenv
